requirements.txt
django ~= 3.0.5
nose   ~= 1.3.7
python-dotenv ~= 0.13.0
daphne ~= 2.5.0
requests ~=	2.8.1
coverage ~= 5.1.0
wheel ~=0.34.2
scrapy ~=2.1.0
scrapy-djangoitem ~=1.1.1
scrapyd ~=1.2.1
python-scrapyd-api ~=2.1.2
mysqlclient ~=1.4.4
djangorestframework ~=3.11.0
